third_party/blink/web_tests/svg/dynamic-updates/SVGFESpotLightElement-dom-y-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y' attribute of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS spotLightElement.getAttribute('y') is "70"
PASS spotLightElement.getAttribute('y') is "-30"
PASS successfullyParsed is true

TEST COMPLETE
